backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.9
pydantic==2.11.3
pydantic-settings==2.8.1
alembic==1.15.2
python-jose==3.4.0
passlib==1.7.4
python-multipart==0.0.20
pytest==8.3.5
python-dotenv==1.1.0
httpx==0.28.1
pillow==11.1.0
